requirements.txt
flask==3.0.2
requests==2.31.0
python-telegram-bot==20.7


